+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 * Moved all database related folders and files in `src/db`.
 * Divided `helpers` files into `functions` and `constants` folders.
 * `checkRouteParameters` helper function renamed into `checkRequestData`.
+* `errors` field from `Error` class renamed to `data`.
 
 ### 📦 Packages